+++ R/layer_hexbin.R
@@ -1,4 +1,3 @@
-
 #' Add a "hexbin" layer to a Bokeh figure
 #' @param fig figure to modify
 #' @param x values or field name of center x coordinates to be binned
@@ -12,6 +11,7 @@
 #' @param line logical - should hexagons have an outline?
 #' @param alpha the alpha transparency of the hexagons between 0 (transparent) and 1 (opaque)
 #' @param hover logical - should a hover tool be added to show the count in each hexagon?
+#' @param visible should the layer be visible?
 #' @examples
 #' \donttest{
 #' figure() %>% ly_hexbin(rnorm(10000), rnorm(10000))
@@ -23,7 +23,7 @@ ly_hexbin <- function(
   style = "colorscale",
   trans = NULL, inv = NULL, lname = NULL,
   palette = "RdYlGn11", line = FALSE, alpha = 1,
-  hover = TRUE
+  hover = TRUE, visible = TRUE
 ) {
 
   args <- sub_names(fig, data,
@@ -39,7 +39,7 @@ ly_hexbin <- function(
       hover,
       line,
       alpha,
-      hover,
+      visible,
       dots = lazy_dots()
     ),
     process_data_and_names = FALSE