+++ R/misc.R
@@ -8,3 +8,16 @@ catjitter <- function(x, factor = 0.8) {
   # validate factor and x
   paste0(x, ":", runif(min = 0.5 * (1 - factor), max = 0.5 + 0.5 * factor, length(x)))
 }
+
+#' Get object ids and types from a figure
+#'
+#' @param fig a figure object
+#' @export
+get_object_refs <- function(fig) {
+  tmp <- lapply(fig$x$spec$model, function(x) {
+    data.frame(id = x$id, type = x$type, stringsAsFactors = FALSE)
+  })
+  res <- do.call(rbind, unname(tmp))
+  res$name <- names(tmp)
+  res
+}