+++ man-roxygen/ex-tap-debug-callback.R
@@ -1,0 +1,5 @@
+\donttest{
+figure() %>%
+  ly_points(1:10, lname = "points") %>%
+  tool_tap(debug_callback("points"), "points")
+}