+++ man-roxygen/par-lnamegroup.R
@@ -1,2 +1,3 @@
 #' @param lname layer name
 #' @param lgroup layer group
+#' @param visible should the layer be visible?